requirements.txt
flask
flask-cors
PyMySQL
cryptography
flask_restful
pillow
